# improve performance for sorting columns in sparse matrix (thereby improving sparse matmul)