This improves performance for sparse matrix multiply (by optimizing the sorting done after the multiply is done). The two main optimizations are:

- use `resize!` instead of `unsafe_wrap`, saves two allocations per column
- Hoist the creation of the `Perm` object, saves at least an allocation per column

Using the benchmarks in https://github.com/JuliaLang/julia/issues/29022

```jl
using SparseArrays
using BenchmarkTools

fun(S) = S*S';
for N = [10000 100000 1000000]
    S = spdiagm(0 => ones(N), 1 => -ones(N-1));
    @show N
    @btime fun($S);
end
```

Before:

```jl
N = 10000
  2.087 ms (30019 allocations: 3.13 MiB)
N = 100000
  23.200 ms (300019 allocations: 31.28 MiB)
N = 1000000
  354.681 ms (3000019 allocations: 312.81 MiB)
```

After

```jl
N = 10000
  536.658 μs (26 allocations: 1.45 MiB)
N = 100000
  6.598 ms (26 allocations: 14.50 MiB)
N = 1000000
  137.312 ms (26 allocations: 144.96 MiB)
```

And benchmarking the code in https://github.com/JuliaLang/julia/issues/29022#issuecomment-430448952

```jl
using MatrixMarket
using SparseArrays
using BenchmarkTools

urls = [
"https://www.cise.ufl.edu/research/sparse/MM/Wissgott/parabolic_fem.tar.gz",
]

for url in urls
    filename = basename(url)
    if !isfile(filename)
        download(url, filename)
        run(`tar -xvf $filename`)
    end
end


fun(S) = S*S';

for url in urls
    path = split(basename(url), ".")[1]
    M = MatrixMarket.mmread(joinpath(path, path*".mtx"))
    println(path)
    @btime fun($M)
end
```

Before

```jl
  571.827 ms (1577502 allocations: 396.69 MiB)
```

After

```jl
  471.738 ms (28 allocations: 308.44 MiB)
```

cc @Alexander-Barth